As path is part of the `Api`, it's methods are included as well. Also added some small steps towards decomposable maps and predicates
