+++ core/src/main/scala/api/package.scala
@@ -4,6 +4,9 @@ import java.nio.file.{ attribute => jnfa }
 import java.util.concurrent.TimeUnit
 
 package object api extends sfs.api.Api {
+
+  implicit class PathOps(path: Path) extends jio.PathOpsInstance(path)
+
   implicit class FileTimeOps(val x: FileTime) extends AnyVal {
     def isOlder(that: FileTime) = (x compareTo that) < 0
     def isNewer(that: FileTime) = (x compareTo that) > 0
